[#120] feat : 데이터 연동 임시
